+++ lib/text_responsive.dart
@@ -99,39 +99,176 @@ class TextResponsiveWidget extends StatelessWidget {
     required double maxWidth,
     required double maxHeight,
     required int maxLines,
-    double initialFontSize =
-        18, // Puede ser un valor predeterminado o basado en la preferencia del usuario
+    double initialFontSize = 18,
   }) {
-    double fontSize =
-        initialFontSize; // Comienza con el tamaño de la fuente inicial
-    late double
-        lastGoodFontSize; // Mantener registro del último tamaño de fuente que encajaba bien
-    bool hasOverflow = true; // Control para el bucle
+    double fontSize = initialFontSize;
+    late double lastGoodFontSize;
+    bool hasOverflow = true;
 
-    while (hasOverflow && fontSize > 1) {
-      // Evita que la fuente sea menor a 1
-      // Configura el TextPainter
+    while (hasOverflow && fontSize > 1.0) {
       final TextPainter textPainter = TextPainter(
         text: TextSpan(text: text, style: style.copyWith(fontSize: fontSize)),
         textDirection: TextDirection.ltr,
         maxLines: maxLines,
       );
 
-      // Realiza el layout del texto
       textPainter.layout(maxWidth: maxWidth);
 
-      // Verifica si hay overflow
       if (textPainter.width <= maxWidth &&
           textPainter.height <= maxHeight &&
           !textPainter.didExceedMaxLines) {
-        lastGoodFontSize =
-            fontSize; // Actualiza el último tamaño de fuente bueno
+        lastGoodFontSize = fontSize;
 
         hasOverflow = false;
       } else {
-        fontSize *= 0.99; // Reduce el tamaño de la fuente
+        fontSize *= 0.99;
       }
     }
-    return lastGoodFontSize; // Devuelve el último tamaño de fuente que encajaba bien
+    return lastGoodFontSize;
+  }
+}
+
+/// A widget that provides responsive text functionality for paragraph text.
+///
+/// This widget wraps a Text widget and automatically resizes the text
+/// to fit its container's width and height while respecting the original
+/// style and structure of the paragraph.
+///
+/// The widget is useful in scenarios where you want to display a paragraph
+/// of text that needs to adapt to the screen size and orientation changes
+/// without manually adjusting the font size.
+///
+/// Example:
+/// ```
+/// ParagraphTextWidget(
+///   'Your paragraph text goes here',
+///   style: TextStyle(fontSize: 20),
+/// ),
+/// ```
+class ParagraphTextWidget extends StatelessWidget {
+  /// Creates a [ParagraphTextWidget].
+  ///
+  /// The [text] parameter must not be null and is the text that the widget displays.
+  /// The [style] parameter is for specifying the style of the text such as fontSize, fontWeight, etc.
+  /// Other parameters like [textAlign], [softWrap], etc., can be set to control how the text behaves.
+  const ParagraphTextWidget(
+    this.text, {
+    super.key,
+    this.style,
+    this.strutStyle,
+    this.textAlign,
+    this.textDirection,
+    this.locale,
+    this.softWrap,
+    this.overflow,
+    this.textScaler,
+    this.maxLines = 10,
+    this.semanticsLabel,
+    this.textWidthBasis,
+    this.selectionColor,
+  });
+
+  final String text;
+  final TextStyle? style;
+  final StrutStyle? strutStyle;
+  final TextAlign? textAlign;
+  final TextDirection? textDirection;
+  final Locale? locale;
+  final bool? softWrap;
+  final TextOverflow? overflow;
+  final TextScaler? textScaler;
+  final int? maxLines;
+  final String? semanticsLabel;
+  final TextWidthBasis? textWidthBasis;
+  final Color? selectionColor;
+
+  @override
+  Widget build(BuildContext context) {
+    final Text child = Text(
+      text,
+      style: style,
+      strutStyle: strutStyle,
+      textAlign: textAlign,
+      textDirection: textDirection,
+      locale: locale,
+      softWrap: softWrap,
+      overflow: overflow,
+      textScaler: textScaler,
+      maxLines: maxLines,
+      semanticsLabel: semanticsLabel,
+      textWidthBasis: textWidthBasis,
+      selectionColor: selectionColor,
+    );
+
+    return TextResponsiveWidget(child: child);
+  }
+}
+
+/// A widget designed to provide responsive text functionality for inline text.
+///
+/// This widget is similar to [ParagraphTextWidget] but is tailored for short,
+/// inline text elements. It resizes text to fit its container's width and height
+/// and maintains the original text style.
+///
+/// This widget is particularly useful for headings, captions, or any short text
+/// that needs to be flexible with its surrounding environment.
+///
+/// Example:
+/// ```
+/// InlineTextWidget(
+///   'Your inline text goes here',
+///   style: TextStyle(fontSize: 20),
+/// ),
+/// ```
+class InlineTextWidget extends StatelessWidget {
+  /// Creates an [InlineTextWidget].
+  ///
+  /// The [text] parameter must not be null and is the text that the widget displays.
+  /// The [style] parameter is for specifying the style of the text.
+  /// Other parameters like [textAlign], [softWrap], etc., can be set to control how the text behaves.
+  const InlineTextWidget(
+    this.text, {
+    super.key,
+    this.style,
+    this.strutStyle,
+    this.textAlign,
+    this.textDirection,
+    this.locale,
+    this.softWrap,
+    this.overflow,
+    this.textScaler,
+    this.semanticsLabel,
+    this.selectionColor,
+  });
+
+  final String text;
+  final TextStyle? style;
+  final StrutStyle? strutStyle;
+  final TextAlign? textAlign;
+  final TextDirection? textDirection;
+  final Locale? locale;
+  final bool? softWrap;
+  final TextOverflow? overflow;
+  final TextScaler? textScaler;
+  final String? semanticsLabel;
+  final Color? selectionColor;
+
+  @override
+  Widget build(BuildContext context) {
+    final Text child = Text(
+      text,
+      style: style,
+      strutStyle: strutStyle,
+      textAlign: textAlign,
+      textDirection: textDirection,
+      locale: locale,
+      softWrap: softWrap,
+      overflow: overflow,
+      textScaler: textScaler,
+      semanticsLabel: semanticsLabel,
+      selectionColor: selectionColor,
+    );
+
+    return TextResponsiveWidget(child: child);
   }
 }